Moved functionality from omtDocumentInformationProvider to omtAvailableObjectsProvider and importMatch
